pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "ultrasound-rl-navigation"
version = "0.1.0"
description = "Ultrasound segmentation + RL navigation (DQN) pipeline"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
authors = [{ name = "Anand Mayank" }]
dependencies = [
  "torch>=1.9.0",
  "torchvision>=0.10.0",
  "numpy>=1.21.0",
  "matplotlib>=3.3.0",
  "Pillow>=8.0.0",
  "tqdm>=4.62.0",
  "opencv-python>=4.5.0",
  "scikit-learn>=1.0.0",
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.23.0",
  "python-multipart>=0.0.6",
  "websockets>=11.0",
]

[project.optional-dependencies]
full = [
  "imageio>=2.32.0",
  "opencv-python-headless>=4.8.1",
  "pandas>=1.5.2",
  "pydicom>=2.4.3",
  "pyk4a>=1.5.0",
  "scikit-image>=0.21.0",
  "scipy>=1.9.3",
  "six>=1.16.0",
  "tabulate>=0.9.0",
  "tensorboard>=2.10.0",
  "tensorboardX>=2.6.2",
  "timm>=0.6.13",
  "torchsummary>=1.5.1",
  "torchaudio>=0.10.0",
  "typing_extensions>=4.4.0",
  "pretrainedmodels>=0.7.4",
]

[project.scripts]
pipeline = "main:run_complete_pipeline"

[tool.setuptools]
packages = ["core", "thyroid_glanod_gym"]
py-modules = [
  "main",
  "train_navigation",
  "train_segmentation",
  "thyroid_glanod_dqn_cjp_train",
  "thyroid_glanod_dqn_predict_patient",
]
include-package-data = false


